backend/pyproject.toml
[project]
name = "audioedit-backend"
version = "0.1.0"
description = "FastAPI backend for audio editing workflows"
requires-python = ">=3.11,<3.13"
dependencies = [
  "demucs>=4.0.1",
  "fastapi>=0.115.0",
  "python-multipart>=0.0.9",
  "torchcodec>=0.11.1",
  "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
  "app",
  "pyproject.toml",
]
